+++ lib/commands/snapshot.php
@@ -9,12 +9,12 @@ use Cranberry\CLI\Command;
 use Cranberry\CLI\Input;
 
 /**
- * @command		Update profile assets from source
- * @desc		update
- * @usage		update [-y|--yes] <app>/<profile>
+ * @command		snapshot
+ * @desc		Update profile assets to mirror live files
+ * @usage		snapshot [-y|--yes] <app>/<profile>
  * @options		-y, --yes
  */
-$command = new Command\Command( 'update', 'Update profile assets from source', function( $query )
+$command = new Command\Command( 'snapshot', 'Update profile assets to mirror live files', function( $query )
 {
 	try
 	{
@@ -22,7 +22,7 @@ $command = new Command\Command( 'update', 'Update profile assets from source', f
 
 		if( is_null( $this->getOptionValue( 'y' ) ) && is_null( $this->getOptionValue( 'yes' ) ) )
 		{
-			$continue = Input::prompt( "Are you sure you want to overwrite the {ul}{$query}{/ul} profile assets? (y/n)", true );
+			$continue = Input::prompt( "Are you sure you want to overwrite the '{$query}' profile assets? (y/n)", true );
 			if( strtolower( $continue ) != 'y' )
 			{
 				exit( 1 );
@@ -44,7 +44,7 @@ $command = new Command\Command( 'update', 'Update profile assets from source', f
 $command->registerOption( 'y' );
 $command->registerOption( 'yes' );
 
-$usage = "update [-y|--yes] <app>/<profile>";
+$usage = "snapshot [-y|--yes] <app>/<profile>";
 $command->setUsage( $usage );
 
 return $command;